Extract password validation logic into a util
